+++ NangPaGo-api/NangPaGo-app/src/main/java/com/mars/app/domain/user_recipe/dto/comment/UserRecipeCommentResponseDto.java
@@ -1,4 +1,4 @@
-package com.mars.app.domain.comment.userRecipe.dto;
+package com.mars.app.domain.user_recipe.dto.comment;
 
 import java.time.LocalDateTime;
