+++ NangPaGo-api/NangPaGo-app/src/main/java/com/mars/app/domain/user_recipe/repository/UserRecipeCommentRepository.java
@@ -1,16 +1,12 @@
-package com.mars.app.domain.comment.userRecipe.repository;
+package com.mars.app.domain.user_recipe.repository;
 
-import com.mars.common.enums.userRecipe.UserRecipeStatus;
 import com.mars.common.model.comment.userRecipe.UserRecipeComment;
-import com.mars.common.model.userRecipe.UserRecipe;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
 
 @Repository
 public interface UserRecipeCommentRepository extends JpaRepository<UserRecipeComment, Long> {
