+++ NangPaGo-api/NangPaGo-app/src/main/java/com/mars/app/domain/user_recipe/repository/UserRecipeRepository.java
@@ -1,4 +1,4 @@
-package com.mars.app.domain.userRecipe.repository;
+package com.mars.app.domain.user_recipe.repository;
 
 import com.mars.common.enums.userRecipe.UserRecipeStatus;
 import com.mars.common.model.userRecipe.UserRecipe;
